fix: import errors due to renaming folders
