+++ src/Router.tsx
@@ -1,5 +1,5 @@
 import Loading from 'components/loading';
-import RequireState from 'components/requireState';
+import RequireState from 'components/require-state';
 import Cute404 from 'pages/page404/Cute404';
 import { lazy } from 'react';
 import { Route, Routes } from 'react-router-dom';
